Use standard socket APIs for non-Linux
